requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
finalizer==1.0.0
fiona==1.9.5
fonttools==4.47.2
geopandas==0.14.2
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
tweepy==4.14.0
twitter-text-parser==3.0.0
tzdata==2023.4
urllib3==2.1.0
